# lab2/contract/products-api/api.yaml
openapi: 3.0.0
info:
  title: soa-products
  description: Product Management API
  version: 1.0.0
...
servers:
  - url: /api
paths:
...
  /products:
    get:
      operationId: getProducts
      tags: [ product ]
      summary: Get all products with optional sorting, filtering, and pagination
      parameters:
        - $ref: 'parameters.yaml#/components/parameters/PageParameter'
        - $ref: 'parameters.yaml#/components/parameters/SizeParameter'
        - $ref: 'parameters.yaml#/components/parameters/SortCriteriaListParameter'
        - $ref: 'parameters.yaml#/components/parameters/SortTypeListParameter'
        - $ref: 'parameters.yaml#/components/parameters/FilterCriteriaListParameter'
        - $ref: 'parameters.yaml#/components/parameters/FilterOperatorsParameter'
        - $ref: 'parameters.yaml#/components/parameters/FilterValuesParameter'
      responses:
        200:
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: 'schemas.yaml#/components/schemas/ProductTo'
        400:
          description: Invalid parameters provided (e.g. negative size, page)
          content:
            application/json:
              schema:
                $ref: 'schemas.yaml#/components/schemas/ApiErrorTo'

    post:
      operationId: createProduct
      tags: [ product ]
      summary: Add a new product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: 'schemas.yaml#/components/schemas/CreateUpdateProductRequestTo'
      responses:
        201:
          description: Product created
          content:
            application/json:
              schema:
                $ref: 'schemas.yaml#/components/schemas/CreateProductResponseTo'
        400:
          description: Invalid request body params (e.g. negative price)
          content:
            application/json:
              schema:
                $ref: 'schemas.yaml#/components/schemas/ApiErrorTo'
        404:
          description: Owner not found
          content:
            application/json:
              schema:
                $ref: 'schemas.yaml#/components/schemas/ApiErrorTo'
  /products/{id}:
    get:
      operationId: getProductById
      tags: [ product ]
      summary: Get a product by ID
      parameters:
        - $ref: 'parameters.yaml#/components/parameters/IdParameter'
      responses:
        200:
          description: Successful response
          content:
            application/json:
              schema:
                $ref: 'schemas.yaml#/components/schemas/ProductTo'
        400:
          description: Invalid id provided
          content:
            application/json:
              schema:
                $ref: 'schemas.yaml#/components/schemas/ApiErrorTo'
        404:
          description: Product not found
          content:
            application/json:
              schema:
                $ref: 'schemas.yaml#/components/schemas/ApiErrorTo'

    put:
      operationId: updateProductById
      tags: [ product ]
      summary: Update a product by ID
      parameters:
        - $ref: 'parameters.yaml#/components/parameters/IdParameter'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: 'schemas.yaml#/components/schemas/CreateUpdateProductRequestTo'
      responses:
        200:
          description: Product updated
        400:
          description: Invalid id provided
          content:
            application/json:
              schema:
                $ref: 'schemas.yaml#/components/schemas/ApiErrorTo'
        404:
          description: Property not found
          content:
            application/json:
              schema:
                $ref: 'schemas.yaml#/components/schemas/UpdateProduct404ResponseTo'

    delete:
      operationId: deleteProductById
      tags: [ product ]
      summary: Delete a product by ID
      parameters:
        - $ref: 'parameters.yaml#/components/parameters/IdParameter'
      responses:
        204:
          description: Product deleted
        400:
          description: Negative id provided
          content:
            application/json:
              schema:
                $ref: 'schemas.yaml#/components/schemas/ApiErrorTo'
        404:
          description: Product not found
          content:
            application/json:
              schema:
                $ref: 'schemas.yaml#/components/schemas/ApiErrorTo'
...